indexed_db: advance scope ids past recovered LevelDBScopes
LevelDBScopes::Initialize() scans existing scope metadata on disk during
startup recovery, but next_scope_id_ remains at its default value.

This lets a newly-created scope reuse an id that already exists on disk,
which can collide with recovered scope metadata or task keys.

Advance next_scope_id_ past each recovered scope id while loading
metadata. Use int64_t for next_scope_id_ to match parsed on-disk scope
ids and LevelDBScope::scope_id(). Reject recovered ids that are negative
or that would overflow when computing the next id.

Bug: 527243514
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7990153
Commit-Queue: Abhishek Shanthkumar <abhishek.shanthkumar@microsoft.com>
Cr-Commit-Position: refs/heads/main@{#1658620}

+++ components/services/storage/indexed_db/scopes/leveldb_scopes.h
@@ -90,7 +90,7 @@ class LevelDBScopes {
 
   // This gets set to |true| when |Initialize()| succeeds.
   bool recovery_finished_ = false;
-  int next_scope_id_ = 0;
+  int64_t next_scope_id_ = 0;
   scoped_refptr<LevelDBState> level_db_;
   // The |lock_manager_| is expected to outlive this class.
   raw_ptr<PartitionedLockManager> lock_manager_;
